refactor(sync): change sync-button to sync-organization-button
