+++ app/views/organizations/_settings_content.html.erb
@@ -8,7 +8,7 @@
       (<%= @organization.repositories.count %>)
       <%= t('messages.settings.repository_title') %>
       <span class="bolder"><%= @github_organization[:login] %></span>
-      <sync-button url="/api/organizations/<%= @organization.id %>/sync" id="<%= @organization.id %>" error_msg="<%= t('messages.settings.error') %>"></sync-button>
+      <sync-organization-button url="/api/organizations/<%= @organization.id %>/sync" id="<%= @organization.id %>" error_msg="<%= t('messages.settings.error') %>"></sync-organization-button>
     </div>
   </div>
   <div class="card-extended__body-list">
